refactored for move to utils folder
