+++ src/components/utils/DateFormater.jsx
@@ -1,6 +1,11 @@
 var moment = require('moment');
 
 const inputStringFormat = 'YYYY-MM';
+const outputStringFormat = 'MMM YYYY';
+
+export function formatDate(date) {
+  return moment(date, inputStringFormat).format(outputStringFormat);
+}
 
 export function formatStartEndDates(startDate, endDate) {
   return `${formatDates(startDate, endDate)} ${formatDuration(startDate, endDate)}`
@@ -10,16 +15,9 @@ function formatDates(startDate, endDate) {
   const startMoment = moment(startDate, inputStringFormat);
   const endMoment = moment(endDate, inputStringFormat);
 
-  let startFormatString;
-  const endFormatString = 'MMM YYYY';
+  let startFormatString = (startMoment.year() === endMoment.year()) ? outputStringFormat : 'MMM';
 
-  if (startMoment.year() === endMoment.year()) {
-    startFormatString = 'MMM';
-  } else {
-    startFormatString = endFormatString;
-  }
-
-  return `${startMoment.format(startFormatString)} - ${endMoment.format(endFormatString)}`
+  return `${startMoment.format(startFormatString)} - ${endMoment.format(outputStringFormat)}`
 }
 
 function formatDuration(startDate, endDate) {
@@ -29,8 +27,9 @@ function formatDuration(startDate, endDate) {
 
   let duration = `${durationMoment.months()} Month${durationMoment.months() > 1 ? 's' : ''}`;
 
-  if (durationMoment.years() !== 0) {
+  if (durationMoment.years() > 0) {
     duration = `${durationMoment.years()} Year${durationMoment.years() > 1 ? 's' : ''} and ${duration}`;
   }
+
   return `(${duration})`;
 }
